// src/api/types/DiscountResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const discountResponseApplicableTypeEnum = {
  COURSE: 'COURSE',
  PACKAGE_ONE_MONTH: 'PACKAGE_ONE_MONTH',
  PACKAGE_THREE_MONTHS: 'PACKAGE_THREE_MONTHS',
  PACKAGE_TWELVE_MONTHS: 'PACKAGE_TWELVE_MONTHS',
  APPOINTMENT: 'APPOINTMENT',
} as const

export type DiscountResponseApplicableTypeEnumKey =
  (typeof discountResponseApplicableTypeEnum)[keyof typeof discountResponseApplicableTypeEnum]

export type DiscountResponse = {
  /**
   * @type integer | undefined, int64
   */
  id?: number
  /**
   * @type string | undefined
   */
  applicableType?: DiscountResponseApplicableTypeEnumKey
  /**
   * @type integer | undefined, int32
   */
  percentage?: number
  /**
   * @type string | undefined, date
   */
  endDate?: string
  /**
   * @type boolean | undefined
   */
  isActive?: boolean
  /**
   * @type string | undefined, date-time
   */
  createdAt?: string
  /**
   * @type string | undefined, date-time
   */
  updatedAt?: string
}